% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_functions.R
\name{additive_regression}
\alias{additive_regression}
\title{Fit sparse BLIN model
D is matrix of lagged Y (covariates)
Y is oservations
X is covariates
for multiple time periods
Returns A, B, beta, Yhat}
\usage{
additive_regression(Y, X, lag, type = "biten", use_cov = TRUE,
  penalty = 1, whichlambda = "min")
}
\description{
Fit sparse BLIN model
D is matrix of lagged Y (covariates)
Y is oservations
X is covariates
for multiple time periods
Returns A, B, beta, Yhat
}
\keyword{internal}
