% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{adiag}
\alias{adiag}
\title{Find diagonal indices of k-mode array, where "diagonal" is i=j for first two indices of input array Y}
\usage{
adiag(Y)
}
\description{
Find diagonal indices of k-mode array, where "diagonal" is i=j for first two indices of input array Y
}
\keyword{internal}
