% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{amprod}
\alias{amprod}
\title{Multiply an array by a matrix along a given mode, by}
\usage{
amprod(A, M, k)
}
\description{
Multiply an array by a matrix along a given mode, by
}
\author{
Peter Hoff

This function multiplies a matricized array by another 
matrix, and then reforms the result into a new array.
}
\keyword{internal}
