% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_ranks}
\alias{check_ranks}
\title{Check inputs for reduced rank fit of BLIN model}
\usage{
check_ranks(rankA, rankB, S, L)
}
\description{
Check inputs for reduced rank fit of BLIN model
}
\keyword{internal}
