% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_unique}
\alias{check_unique}
\title{Check whether there is a unique solution to the BLIN model}
\usage{
check_unique(S, L, tmax)
}
\description{
Check whether there is a unique solution to the BLIN model
}
\keyword{internal}
