% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blinfit.R
\name{coef.blin}
\alias{coef.blin}
\title{Coef S3 generic for class blin}
\usage{
\method{coef}{blin}(object, whichcoef = NULL, ...)
}
\arguments{
\item{object}{blin object}

\item{whichcoef}{optional string (or NULL) indicating which coefficient to retrun, i.e. A, B, beta, or diagAB. If NULL, returns list of all coefficients.}

\item{...}{ignored}
}
\description{
Coef S3 generic for class blin
}
