% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_functions.R
\name{fit_MLE_array}
\alias{fit_MLE_array}
\title{Fit reduced rank BLIN model
D is matrix of lagged Y (covariates)
Y is oservations
X is covariates
for multiple time periods
Returns A, B, beta, betaOLS, Yhat, 2LL, 2LLinit}
\usage{
fit_MLE_array(Y, D, X, lag, rankA = 1, rankB = 1, verbose = FALSE,
  printout = 1, tol = 1e-08, init = "I", sigma_init = 1,
  use_cov = TRUE, maxit = 1000, randseed = NA, Yold = NULL)
}
\description{
Fit reduced rank BLIN model
D is matrix of lagged Y (covariates)
Y is oservations
X is covariates
for multiple time periods
Returns A, B, beta, betaOLS, Yhat, 2LL, 2LLinit
}
\keyword{internal}
