% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blinfit.R
\name{model.matrix.blin}
\alias{model.matrix.blin}
\title{model.matrix S3 generic for class blin}
\usage{
\method{model.matrix}{blin}(object, ...)
}
\arguments{
\item{object}{blin object}

\item{...}{ignored}
}
\description{
model.matrix S3 generic for class blin
}
