% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blinfit.R
\name{plot.blin}
\alias{plot.blin}
\title{Plot S3 generic for class blin}
\usage{
\method{plot}{blin}(x, ...)
}
\arguments{
\item{x}{blin object}

\item{...}{ignored}
}
\description{
Plot S3 generic for class blin
}
