% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blinfit.R
\name{print.blin}
\alias{print.blin}
\title{Print S3 generic for class blin}
\usage{
\method{print}{blin}(x, hn = 10, ...)
}
\arguments{
\item{x}{blin object}

\item{hn}{optional numeric length of each coefficient printed}

\item{...}{ignored}
}
\description{
Print S3 generic for class blin
}
