% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blinfit.R
\name{print.summary.blin}
\alias{print.summary.blin}
\title{Print S3 generic for class summary.blin}
\usage{
\method{print}{summary.blin}(x, hn = 10, ...)
}
\arguments{
\item{x}{summary.blin object}

\item{hn}{optional numeric length of each coefficient printed}

\item{...}{ignored}
}
\description{
Print S3 generic for class summary.blin
}
