% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_functions.R
\name{update_MLE_additive}
\alias{update_MLE_additive}
\title{Update for block coordinate descent of full BLIN model
D is matrix of lagged Y (covariates)
Y is observations}
\usage{
update_MLE_additive(A, B, D, DDT, DTD, DYT, DTY, type = "biten")
}
\description{
Update for block coordinate descent of full BLIN model
D is matrix of lagged Y (covariates)
Y is observations
}
\keyword{internal}
