% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_functions.R
\name{update_MLE_asymmetric}
\alias{update_MLE_asymmetric}
\title{D is matrix of lagged Y (covariates)
Y is observations}
\usage{
update_MLE_asymmetric(D, U, V, W, Z, DDT, DTD, DYT, DTY)
}
\description{
D is matrix of lagged Y (covariates)
Y is observations
}
\keyword{internal}
