% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blinfit.R
\name{vcov.blin}
\alias{vcov.blin}
\title{vcov S3 generic for class blin}
\usage{
\method{vcov}{blin}(object, ...)
}
\arguments{
\item{object}{blin object}

\item{...}{ignored}
}
\description{
vcov S3 generic for class blin
}
